requirements.txt
Flask==0.12.2
line-bot-sdk==1.5.0
future==0.18.2
requests==2.18.4
click==6.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe
Werkzeug==0.14.1

mysqlclient==1.3.12
PyMySQL==0.10.1
mysql-connector-python==8.0.22
urllib3==1.25.11
beautifulsoup4==4.9.3